refactor: keep power quote modules bounded
